refactor: share stats widget story and test helpers
