+++ projects/packages/premium-analytics/changelog/wooa7s-1755-story-test-helpers
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Widgets: Keep Storybook rendering and test harnesses consistent across Stats widgets.
