+++ projects/packages/premium-analytics/widgets/email-breakdown/__tests__/email-breakdown.test.tsx
@@ -36,9 +36,7 @@ jest.mock( '@jetpack-premium-analytics/widgets-toolkit', () => ( {
 
 // WidgetRoot reads URL search params as a fallback for report params; outside
 // a matched route the real hook warns and throws.
-jest.mock( '@wordpress/route', () => ( {
-	useSearch: () => ( {} ),
-} ) );
+jest.mock( '@wordpress/route', () => jest.requireActual( '../../test-utils' ).mockWordPressRoute );
 
 const mockApiFetch = apiFetch as unknown as jest.Mock;
