+++ projects/packages/premium-analytics/widgets/plan-usage/stories/plan-usage-widget.stories.tsx
@@ -27,9 +27,10 @@ import {
 	widgetDashboardWithWidgetArgTypes,
 	type WidgetDashboardWithWidgetControls,
 } from '../../stories/widget-dashboard-with-widget';
+import { withWidgetCanvas } from '../../stories/with-widget-canvas';
 import PlanUsageRender from '../render';
 import widgetDefinition from '../widget';
-import type { Decorator, Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react';
 import type { WidgetRenderProps } from '@wordpress/widget-primitives';
 import type { ComponentProps, ComponentType } from 'react';
 
@@ -120,13 +121,6 @@ function renderPlanUsage( { withComparison }: PlanUsageStoryControls ) {
 		<PlanUsageRender attributes={ { reportParams: getDefaultQueryParams( withComparison ) } } />
 	);
 }
-
-// Close-up canvas so the gauge fills the frame outside the dashboard grid.
-const withWidgetCanvas: Decorator = Story => (
-	<div style={ { width: '100%', height: '360px' } }>
-		<Story />
-	</div>
-);
 
 const meta = {
 	title: 'Packages/Premium Analytics/Widgets/PlanUsage',
