+++ projects/packages/premium-analytics/widgets/post-comments/stories/post-comments-widget.stories.tsx
@@ -12,9 +12,10 @@ import {
 	widgetDashboardWithWidgetArgTypes,
 	type WidgetDashboardWithWidgetControls,
 } from '../../stories/widget-dashboard-with-widget';
+import { withWidgetCanvas } from '../../stories/with-widget-canvas';
 import PostCommentsRender from '../render';
 import widgetDefinition from '../widget';
-import type { Decorator, Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react';
 import type { WidgetRenderProps } from '@wordpress/widget-primitives';
 import type { ComponentProps, ComponentType } from 'react';
 
@@ -41,12 +42,6 @@ function getPostCommentsAttributes( {
 function renderPostComments( controls: PostCommentsStoryControls ) {
 	return <PostCommentsRender attributes={ getPostCommentsAttributes( controls ) } />;
 }
-
-const withWidgetCanvas: Decorator = Story => (
-	<div style={ { width: '360px', height: '480px' } }>
-		<Story />
-	</div>
-);
 
 const meta = {
 	title: 'Packages/Premium Analytics/Widgets/PostComments',
