+++ projects/packages/premium-analytics/widgets/post-likes/stories/post-likes-widget.stories.tsx
@@ -23,9 +23,10 @@ import {
 	widgetDashboardWithWidgetArgTypes,
 	type WidgetDashboardWithWidgetControls,
 } from '../../stories/widget-dashboard-with-widget';
+import { withWidgetCanvas } from '../../stories/with-widget-canvas';
 import PostLikesRender from '../render';
 import widgetDefinition from '../widget';
-import type { Decorator, Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from '@storybook/react';
 import type { WidgetRenderProps } from '@wordpress/widget-primitives';
 import type { ComponentProps, ComponentType } from 'react';
 
@@ -68,13 +69,6 @@ function getPostLikesAttributes( {
 function renderPostLikes( controls: PostLikesStoryControls ) {
 	return <PostLikesRender attributes={ getPostLikesAttributes( controls ) } />;
 }
-
-// Close-up canvas sized to the roster outside the dashboard grid.
-const withWidgetCanvas: Decorator = Story => (
-	<div style={ { width: '360px', height: '480px' } }>
-		<Story />
-	</div>
-);
 
 const meta = {
 	title: 'Packages/Premium Analytics/Widgets/PostLikes',
